Ball and BallRender into different modules
- Ball no longer has render_obj attr
- BallRender has ball attr
